添加构建Lua解释器Part05的windows工程

+++ windows/project/part05/test/p2_test.c
@@ -33,7 +33,7 @@ void p2_test_main() {
             printf("timestamp:%d totalbytes:%f kb \n", (int)time(NULL), (float)(g->totalbytes + g->GCdebt) / 1024.0f);
         }
    }
-   end_time = time(NULL);
+   end_time = (int)time(NULL);
    printf("finish test start_time:%d end_time:%d max_bytes:%f kb \n", start_time, end_time, (float)max_bytes / 1024.0f);
 
    luaL_close(L);
